- bug #534 fixed. - some minor refactoring
