+++ openchain/crypto/node_crypto.go
@@ -42,6 +42,11 @@ func (node *nodeImpl) initCryptoEngine() error {
  		return err
  	}
 
+	// Load tls certificate
+	if err := node.loadTLSCertificate(); err != nil {
+		return err
+	}
+
 	node.log.Info("Initializing node crypto engine...done!")
 
 	return nil
