+++ openchain/crypto/node_eca.go
@@ -33,6 +33,7 @@ import (
 	"github.com/openblockchain/obc-peer/openchain/crypto/utils"
 	"golang.org/x/net/context"
 	"google.golang.org/grpc"
+	"io/ioutil"
 )
 
 func (node *nodeImpl) retrieveECACertsChain(userID string) error {
@@ -60,6 +61,122 @@ func (node *nodeImpl) retrieveECACertsChain(userID string) error {
 		node.log.Error("Failed storing eca certificate [%s].", err.Error())
 		return err
 	}
+
+	return nil
+}
+
+func (node *nodeImpl) retrieveEnrollmentData(enrollID, enrollPWD string) error {
+	key, enrollCertRaw, enrollChainKey, err := node.getEnrollmentCertificateFromECA(enrollID, enrollPWD)
+	if err != nil {
+		node.log.Error("Failed getting enrollment certificate [id=%s]: [%s]", enrollID, err)
+
+		return err
+	}
+	node.log.Debug("Enrollment certificate [%s].", utils.EncodeBase64(enrollCertRaw))
+
+	node.log.Debug("Storing enrollment data for user [%s]...", enrollID)
+
+	// Store enrollment id
+	err = ioutil.WriteFile(node.conf.getEnrollmentIDPath(), []byte(enrollID), 0700)
+	if err != nil {
+		node.log.Error("Failed storing enrollment certificate [id=%s]: [%s]", enrollID, err)
+		return err
+	}
+
+	// Store enrollment key
+	if err := node.ks.storePrivateKey(node.conf.getEnrollmentKeyFilename(), key); err != nil {
+		node.log.Error("Failed storing enrollment key [id=%s]: [%s]", enrollID, err)
+		return err
+	}
+
+	// Store enrollment cert
+	if err := node.ks.storeCert(node.conf.getEnrollmentCertFilename(), enrollCertRaw); err != nil {
+		node.log.Error("Failed storing enrollment certificate [id=%s]: [%s]", enrollID, err)
+		return err
+	}
+
+	// Store enrollment chain key
+	if err := node.ks.storeKey(node.conf.getEnrollmentChainKeyFilename(), enrollChainKey); err != nil {
+		node.log.Error("Failed storing enrollment chain key [id=%s]: [%s]", enrollID, err)
+		return err
+	}
+
+	return nil
+}
+
+func (node *nodeImpl) loadEnrollmentKey() error {
+	node.log.Debug("Loading enrollment key...")
+
+	enrollPrivKey, err := node.ks.loadPrivateKey(node.conf.getEnrollmentKeyFilename())
+	if err != nil {
+		node.log.Error("Failed loading enrollment private key [%s].", err.Error())
+
+		return err
+	}
+
+	node.enrollPrivKey = enrollPrivKey.(*ecdsa.PrivateKey)
+
+	return nil
+}
+
+func (node *nodeImpl) loadEnrollmentCertificate() error {
+	node.log.Debug("Loading enrollment certificate...")
+
+	cert, der, err := node.ks.loadCertX509AndDer(node.conf.getEnrollmentCertFilename())
+	if err != nil {
+		node.log.Error("Failed parsing enrollment certificate [%s].", err.Error())
+
+		return err
+	}
+	node.enrollCert = cert
+
+	// TODO: move this to retrieve
+	pk := node.enrollCert.PublicKey.(*ecdsa.PublicKey)
+	err = utils.VerifySignCapability(node.enrollPrivKey, pk)
+	if err != nil {
+		node.log.Error("Failed checking enrollment certificate against enrollment key [%s].", err.Error())
+
+		return err
+	}
+
+	// Set node ID
+	node.id = utils.Hash(der)
+	node.log.Debug("Setting id to [%s].", utils.EncodeBase64(node.id))
+
+	// Set eCertHash
+	node.enrollCertHash = utils.Hash(der)
+	node.log.Debug("Setting enrollCertHash to [%s].", utils.EncodeBase64(node.enrollCertHash))
+
+	return nil
+}
+
+func (node *nodeImpl) loadEnrollmentID() error {
+	node.log.Debug("Loading enrollment id at [%s]...", node.conf.getEnrollmentIDPath())
+
+	enrollID, err := ioutil.ReadFile(node.conf.getEnrollmentIDPath())
+	if err != nil {
+		node.log.Error("Failed loading enrollment id [%s].", err.Error())
+
+		return err
+	}
+
+	// Set enrollment ID
+	node.enrollID = string(enrollID)
+	node.log.Debug("Setting enrollment id to [%s].", node.enrollID)
+
+	return nil
+}
+
+func (node *nodeImpl) loadEnrollmentChainKey() error {
+	node.log.Debug("Loading enrollment chain key...")
+
+	enrollChainKey, err := node.ks.loadKey(node.conf.getEnrollmentChainKeyFilename())
+	if err != nil {
+		node.log.Error("Failed loading enrollment chain key [%s].", err.Error())
+
+		return err
+	}
+	node.enrollChainKey = enrollChainKey
 
 	return nil
 }
